Rename and remove features
